+++ src/m365/pp/commands/dataverse/dataverse-table-row-remove.spec.ts
@@ -1,0 +1,244 @@
+import * as assert from 'assert';
+import * as sinon from 'sinon';
+import auth from '../../../../Auth';
+import { Cli } from '../../../../cli/Cli';
+import { CommandInfo } from '../../../../cli/CommandInfo';
+import { Logger } from '../../../../cli/Logger';
+import Command, { CommandError } from '../../../../Command';
+import request from '../../../../request';
+import { telemetry } from '../../../../telemetry';
+import { pid } from '../../../../utils/pid';
+import { powerPlatform } from '../../../../utils/powerPlatform';
+import { sinonUtil } from '../../../../utils/sinonUtil';
+import commands from '../../commands';
+const command: Command = require('./dataverse-table-row-remove');
+
+describe(commands.DATAVERSE_TABLE_ROW_REMOVE, () => {
+  let commandInfo: CommandInfo;
+  //#region Mocked Responses
+  const validEnvironment = '4be50206-9576-4237-8b17-38d8aadfaa36';
+  const validId = '3a081d91-5ea8-40a7-8ac9-abbaa3fcb893';
+  const validTableName = 'DataverseTable';
+  const validEntitySetName = 'cr6c3_dataversetables';
+  const envUrl = "https://contoso-dev.api.crm4.dynamics.com";
+  const tableResponse = {
+    EntitySetName: 'cr6c3_dataversetables'
+  };
+  //#endregion
+
+  let log: string[];
+  let logger: Logger;
+  let promptOptions: any;
+  let loggerLogToStderrSpy: sinon.SinonSpy;
+
+  before(() => {
+    sinon.stub(auth, 'restoreAuth').callsFake(() => Promise.resolve());
+    sinon.stub(telemetry, 'trackEvent').callsFake(() => { });
+    sinon.stub(pid, 'getProcessName').callsFake(() => '');
+    auth.service.connected = true;
+    commandInfo = Cli.getCommandInfo(command);
+  });
+
+  beforeEach(() => {
+    log = [];
+    logger = {
+      log: (msg: string) => {
+        log.push(msg);
+      },
+      logRaw: (msg: string) => {
+        log.push(msg);
+      },
+      logToStderr: (msg: string) => {
+        log.push(msg);
+      }
+    };
+    loggerLogToStderrSpy = sinon.spy(logger, 'logToStderr');
+    sinon.stub(Cli, 'prompt').callsFake(async (options: any) => {
+      promptOptions = options;
+      return { continue: false };
+    });
+    promptOptions = undefined;
+  });
+
+  afterEach(() => {
+    sinonUtil.restore([
+      request.get,
+      request.delete,
+      powerPlatform.getDynamicsInstanceApiUrl,
+      Cli.prompt,
+      Cli.executeCommandWithOutput
+    ]);
+  });
+
+  after(() => {
+    sinonUtil.restore([
+      auth.restoreAuth,
+      telemetry.trackEvent,
+      pid.getProcessName
+    ]);
+    auth.service.connected = false;
+  });
+
+  it('has correct name', () => {
+    assert.strictEqual(command.name, commands.DATAVERSE_TABLE_ROW_REMOVE);
+  });
+
+  it('has a description', () => {
+    assert.notStrictEqual(command.description, null);
+  });
+
+  it('fails validation if id is not a valid guid.', async () => {
+    const actual = await command.validate({
+      options: {
+        environment: validEnvironment,
+        id: 'Invalid GUID',
+        tableName: validTableName
+      }
+    }, commandInfo);
+    assert.notStrictEqual(actual, true);
+  });
+
+  it('passes validation if required options specified (tableName)', async () => {
+    const actual = await command.validate({ options: { environment: validEnvironment, tableName: validTableName, id: validId } }, commandInfo);
+    assert.strictEqual(actual, true);
+  });
+
+  it('passes validation if required options specified (entitySetName)', async () => {
+    const actual = await command.validate({ options: { environment: validEnvironment, entitySetName: validEntitySetName, id: validId } }, commandInfo);
+    assert.strictEqual(actual, true);
+  });
+
+  it('prompts before removing the specified row from a dataverse table owned by the currently signed-in user when confirm option not passed', async () => {
+    sinon.stub(powerPlatform, 'getDynamicsInstanceApiUrl').callsFake(async () => envUrl);
+
+    await command.action(logger, {
+      options: {
+        environment: validEnvironment,
+        id: validId
+      }
+    });
+    let promptIssued = false;
+
+    if (promptOptions && promptOptions.type === 'confirm') {
+      promptIssued = true;
+    }
+
+    assert(promptIssued);
+  });
+
+  it('aborts removing the specified row from a dataverse table owned by the currently signed-in user when confirm option not passed and prompt not confirmed', async () => {
+    const postSpy = sinon.spy(request, 'delete');
+    sinonUtil.restore(Cli.prompt);
+    sinon.stub(Cli, 'prompt').callsFake(async () => (
+      { continue: false }
+    ));
+    await command.action(logger, {
+      options: {
+        environment: validEnvironment,
+        id: validId
+      }
+    });
+    assert(postSpy.notCalled);
+  });
+
+  it('removes the specified row according to the entitySetName parameter from a dataverse table owned by the currently signed-in user when prompt confirmed', async () => {
+    sinon.stub(powerPlatform, 'getDynamicsInstanceApiUrl').callsFake(async () => envUrl);
+
+    sinon.stub(request, 'delete').callsFake(async (opts) => {
+      if (opts.url === `https://contoso-dev.api.crm4.dynamics.com/api/data/v9.1/${validEntitySetName}(${validId})`) {
+        return;
+      }
+
+      throw 'Invalid request';
+    });
+
+    sinonUtil.restore(Cli.prompt);
+    sinon.stub(Cli, 'prompt').callsFake(async () => (
+      { continue: true }
+    ));
+    await command.action(logger, {
+      options: {
+        debug: true,
+        environment: validEnvironment,
+        id: validId,
+        entitySetName: validEntitySetName
+      }
+    });
+    assert(loggerLogToStderrSpy.called);
+  });
+
+  it('removes the specified row according to the tableName parameter from a dataverse table with the entitySetName parameter without confirmation prompt', async () => {
+    sinon.stub(powerPlatform, 'getDynamicsInstanceApiUrl').callsFake(async () => envUrl);
+
+    sinon.stub(request, 'get').callsFake(async opts => {
+      if ((opts.url === `https://contoso-dev.api.crm4.dynamics.com/api/data/v9.0/EntityDefinitions(LogicalName='${validTableName}')?$select=EntitySetName`)) {
+        if ((opts.headers?.accept as string)?.indexOf('application/json') === 0) {
+          return tableResponse;
+        }
+      }
+
+      throw 'Invalid request';
+    });
+
+    sinon.stub(request, 'delete').callsFake(async (opts) => {
+      if (opts.url === `https://contoso-dev.api.crm4.dynamics.com/api/data/v9.1/${validEntitySetName}(${validId})`) {
+        return;
+      }
+
+      throw 'Invalid request';
+    });
+
+    await command.action(logger, {
+      options: {
+        debug: true,
+        environment: validEnvironment,
+        id: validId,
+        tableName: validTableName,
+        confirm: true
+      }
+    });
+    assert(loggerLogToStderrSpy.called);
+  });
+
+  it('correctly handles API OData error', async () => {
+    const errorMessage = 'Something went wrong';
+
+    sinon.stub(powerPlatform, 'getDynamicsInstanceApiUrl').callsFake(async () => envUrl);
+
+    sinon.stub(request, 'delete').callsFake(async () => { throw { error: { error: { message: errorMessage } } }; });
+
+    await assert.rejects(command.action(logger, {
+      options: {
+        debug: true,
+        environment: validEnvironment,
+        id: validId,
+        confirm: true,
+        entitySetName: validEntitySetName
+      }
+    }), new CommandError(errorMessage));
+  });
+
+  it('removes dataverse table row with the entitySetName parameter and without confirmation', async () => {
+    sinon.stub(powerPlatform, 'getDynamicsInstanceApiUrl').callsFake(async () => envUrl);
+
+    sinon.stub(request, 'delete').callsFake(async (opts) => {
+      if (opts.url === `https://contoso-dev.api.crm4.dynamics.com/api/data/v9.1/${validEntitySetName}(${validId})`) {
+        return;
+      }
+
+      throw 'Invalid request';
+    });
+
+    await command.action(logger, {
+      options: {
+        debug: true,
+        environment: validEnvironment,
+        id: validId,
+        entitySetName: validEntitySetName,
+        confirm: true
+      }
+    });
+
+    assert(loggerLogToStderrSpy.called);
+  });
+});